improved api calls for the cryto exchanges
improved the api calls to the alpha vantage to retrive the exchange
values. no longer has an api call for each cryto now only uses two as
the api only allows for 5 call per min and the application will have 6
or more.Started the api call for the coin page after the coin-list
conponant but still needs more work.