bare bones of location started
location componant has been started and is functional. at the moment it
gets the users location and displays it on a map. this code was got from
moodle page and in further development will allow for the user to pick a
location if they do not wish to use their gps. once a location is picked
it will show the banks in the area. this will be done using a google api
and the key is already gotten.