bare bones of location html page
the location componant has been started and is currently a work in
progress. the code was taken form moodle to see it in action and to
judge how to mod it to the app needs. this will be improved as time goes
on to have a drop down with locations and will display the map of banks
in the area on either this page or on banks commponant