improvments made to work with the improved api calls
with the new api calls on the coin service changes were made to work
with these changes such as passing the varrible into the the two
functions instead of having a different function for each. reason for
two funtions was to work with the two new api calls that are using
different apikeys.