reduced the data being pulled and link to coin
the data being pulled from the api was change to the exact info need for
this page in the correct order (name, stock code and price). code has
been tested and is working. it in commented out to reduce api calls made
as max is 500 per day per apikey. each cryto coin was placed into a div
to seperate them and to start making links to the coin page using
routerlink. this will be improved to cards and have an image of each
coin with annimation. not fully finished needs to add ngif to correct
error messages