started work on coin componant linked to coin-list
began work connecting coin to coin-list. this is so when a coin div is
clicked on in coin-list will bring it to that coin page using the coins
code from exhange api. work has also been started on getting data from
the api on its stocks in each week. this data will be passed into a
chart in the future (chartjs)