Improvement to api calls and chart api call added
the api calls to the whole app was improved. the cryto coin exchange
api calls is now only 2 instead of having a seperate call for each. the
reason for having two for this is cause the api only allows for 5 api
calls per min and the app may have more than 4 cryto coins. the api call
to collect the data for the stock of the coin has been added to populate
the chartjs.