started the location component
the bare bones of the location componant has been started. the code from
moodle has been added so for now it get the users location nad displays
on a map. this will be change in the future to display the user location
with the closet banks near them. it will also have a drop down menu
listing locations if user does now wish to use their gps