new layout and new cryto coins added
a new layout has been added to start the styling stage of the app.
bootstrap will be used to help style. each coin has been placed in a div
container and than a row div. links have been added to the container div
to bring the user to the stock chart of that coin. more coins have been
added that are popular in this martket but is subjected to change with
more or less  as time goes on.