change sytle to scss and api stock data for soins
the style has been change to scss instead of sass. the data needed for
the cryto coin stock data is no working thanks to the help of Shane.
it gets the data from the api object and turns it into an array that can
then be passed into chartjs. it currently only gets the high and low of
the cryto coin but this will change in the future to get the open and
close of the market. the object now only get the first 30 children using
slice as it was getting 1886 back. improvements to the charts design
will be done in the future.