added basic styling of the component
starting the sytling of this component to imporve the visional of the
app. container div was used to add a background colour of grey with 80%
opacity and a height. inside that div is a row div with a flex box to
evenly seperate the internal parts. improved will be made in the future