basic header component in place
a basic header component was made to better understand how to design the
app. after some research it has been decided for now that this will only
be used as a title of the app and maybe a back button on pages. the main
navigation will be in the footer with as this in the current style seen
on mobile apps