styling has begun
began styling of this component to try decide how to make it look. row
div was used to make text linear with a flexbox to devide up the text
evenly. changes made to the table such as adding padding but more will
be done in future. changes made to the H tags to put size and color on
each. hr tag added color and witdh to 90%