imporve data format from api and to the chart
first change was to the table to remove the open and close data from it
so just displays the high and lows of each day. no longer displays 30
days of data now down to 7. style changes to chart was also done e.g
label colors and lines. better format the data being pulled in from to
allow the currency pipe to work on the html page. more improvements will
be made to do the same to the data in the chart as problems were being
had for this.