header, footer and body
added a gobal header footer and body div to the app. the reason for this
was to to allow the header and footer to be fixed to the top and bottom
of the app with set height and to allow the body which is the other
components to move in between without them covering any of their views.
also added a font size and border box to the html to ensure rem will be
accrute on all components