TranslateModule, TranslateLoader and TranslateHttp
importing TranslateModule, TranslateLoader and TranslateHttpLoader to
the app. these will be used for the translation of english to french
within the app. added to functions in this file as well that allow for
the translation to take effect without having to load the whole page.
all this will be connected to the i18n in the assets folder thats sorts
the transtaltion