getting the user location and drop down menu
clean up the code to allow it to be easier to read. created the data for
a dropdown menu if the user didnt want to use gps. it currently has 4
towns in it that also get session stored for easy access. improved the
geolocation code to suit the api better and session store it. This was
down with the help of a student called stefan