began styling of the home page
first bit of styleing has been done on the home page. this has included
the font sizes and colours. div paramaters such as margins width and
border radis. have created class for the img going to be display to hep
decide on height. this will all be change as styling continues