cleaned up the code, and change the layout
cleaned up the code to make it easier to read. change the layout by
changing the size and padding of the h tags, added a flex box to the row
of details about that crypto stock at the bottom of the page. changed
the sizing of the line graph and the graph text sizes