rearranged the header
removed the flags for translation and changed it into a dropdown menu.
added a back button connected to a arrow icon created in photoshop and
changed the logo of the app. removed the title of the app. cleaned up
the code to make it easier to read and made it into a flexbox