improved the tsyling of the header
made the header into a flexbox. changed the sizing of the logo. styled
the dropdown menu  and added a style to the back arrow icon with set
style padding and margin. changed the colour of the header to suit the
new style of the app. cleaned up the code to make it easier to read